ged: Add some extra comments to explain MSI logic
One of the main goal for NEMU is to remove all unneeded legacy
components. In case of ACPI GED device, the goal is to move away
from its dependency on IOAPIC. This can be achieved by generating
interrupts through MSI messages, which means the interrupts will
be directly delivered to the local APIC.

The MSI support from GED assumes the GED driver from the guest OS
has been modified accordingly, by adding the support for the new
ACPI resource MessageSignalledInterrupt. This resource will allow
the guest OS to pass down to the hypervisor the information about
where to send MSI messages, and what to include in the content of
the message, relying on the _SRS method.

In more details, the guest OS will provide the 64 bits address where
to write the MSI message. The vector number and the flags related to
the delivery modes will be passed through a double word (32 bits).

There is one more parameter that we haven't mentioned yet, that can
be used as an identifier to match a set of information (address and
data) with each MessageSignalledInterrupt() entry listed from _CRS
resources of the GED device. This is useful for potential
implementations of GED wanting the MSI support without relying on
any multiplexing.

As a summary, here is what the recent code adding the MSI support
for GED actually does. It defines a new ACPI resource type, used
from _CRS, _PRS and _SRS methods, part of the GED device definition.
Then it creates new IO ports that will be later used to retrieve the
information written by the guest OS through the evaluation of _SRS
ACPI method. Once MSI information are known by the platform, any
device hotplug action (CPU/Memory/PCI) will result in the hypervisor
creating and sending a new MSI message directly to the virtual local
APIC.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>